+++ doc/CHANGES.rst
@@ -12,6 +12,13 @@
 CHANGELOG
 =========
 
+Future
+------
+
+* [REF] adapt to AnyBlok 0.9.0 and AnyBlok / Pyramid 0.7.0
+* [REF] replace set/get callable by plugin get_db_name
+* [REF] Update doc
+
 O.2.0 (2016-06-22)
 ------------------